+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## change log ##
 
 * 2017-02-03:  3.1.8
+  - improve reset-password and password-reset html pages
   - regenerate Cfg with tscfg 0.8.0.  Use json4s to log out the configuration.
   
 * 2017-02-01:  3.1.7